Finished first implementation of TIRBM

+++ RBMCPU/TIRBM.h
@@ -10,7 +10,7 @@ using namespace std;
 class TIRBM
 {
 private:
-	vector<Symmetry<int>> symmetries;
+	vector<Symmetry<double>> symmetries;
 	//weights
 	vector<vector<double>> wij;
 	//last changes for the weights (e.g. used for momentum)
@@ -38,11 +38,11 @@ private:
 	//sample from unifrom dist
 	double uniform(double min, double max);
 	//positive step
-	void sample_h_given_v(vector<int> &vis_src, vector<vector<double>> &hid_target, vector<int> &hid_target_sample, vector<int> &max_pooled_s);
+	void sample_h_given_v(vector<double> &vis_src, vector<vector<double>> &hid_target, vector<double> &hid_target_sample, vector<int> &max_pooled_s);
 	//negative step
-	void sample_v_given_h(vector<int> &hid_src, vector<double> &vis_target, vector<int> &vis_target_sample, vector<int> &max_pooled_s);
+	void sample_v_given_h(vector<double> &hid_src, vector<double> &vis_target, vector<double> &vis_target_sample, vector<int> &max_pooled_s);
 	//approximate the gradient for the log likelyhood
-	double contrastive_divergence(vector<vector<int>> &input, int cdK, int batchSize);
+	double contrastive_divergence(vector<vector<double>> &input, int cdK, int batchSize);
 	//choose the highest value from array
 	int max_pool(vector<double> hid_fixedj);
 	ActivationFunction actFun;
